refactor : repository
member 와 problem 에 관한 repo 를 폴더별로 구분하였습니다.
