+++ poj/src/main/java/com/poj/factory/member/MemberFactory.java
@@ -3,8 +3,8 @@ package com.poj.factory.member;
 import com.poj.dto.auth.SignupRequest;
 import com.poj.entity.member.EAuthority;
 import com.poj.entity.member.Member;
-import com.poj.repository.AuthorityRepository;
-import com.poj.repository.MemberRepository;
+import com.poj.repository.member.AuthorityRepository;
+import com.poj.repository.member.MemberRepository;
 import com.poj.validation.auth.AuthValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
