+++ poj/src/main/java/com/poj/repository/problem/ProblemRepository.java
@@ -1,4 +1,4 @@
-package com.poj.repository;
+package com.poj.repository.problem;
 
 import com.poj.entity.problem.Problem;
 import org.springframework.data.jpa.repository.JpaRepository;
